services/admin_panel/requirements.txt
fastapi>=0.128.0
uvicorn>=0.30.0
sqladmin>=0.20.1
sqlalchemy>=2.0.45
psycopg[binary]>=3.2.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
jinja2>=3.1.4
python-multipart>=0.0.9
itsdangerous>=2.2.0
